Consolidate viewer lifecycle into ViewerContext
